2023-08-24 A windows console terminalban javitva az elromlo szelso oszlop. A terminal hibaja miatt (?) a maxcol()-1-re vegzodo sorok kiirasakor elromik a maxcol() pozicio. Ezert az ilyen sorokat ket reszben irjuk: eloszor maxcol()-2-ig, utana maxcol()-1-tol maxcol()-ig.

+++ ccclib/src/sckutil.cpp
@@ -39,7 +39,7 @@ int socket_wsastartup()
             WSADATA Data;
             if( WSAStartup(MAKEWORD(1,1),&Data) ) //0:success
             {
-                fprintf(stderr,"WSAStartup failed (%d)\n",GetLastError());
+                fprintf(stderr,"WSAStartup failed (%d)\n",(int)GetLastError());
             }
         }
     #endif
